requirements.txt
flask==2.3.2
ollama==0.4.5
llama-index
llama-agents 
llama-index-vector-stores-chroma
llama-index-readers-file
llama-index-llms-ollama
llama-index-embeddings-huggingface
llama-index-embeddings-ollama
llama-index-utils-workflow
chromadb
huggingface_hub
llama-index-llms-huggingface
llama-index-llms-huggingface-api
transformers
torch
typing
nest_asyncio
docx2txt
pypdf
python-dotenv
torch 
torchvision 
torchaudio
langchain
langchain-ollama

